Refactor 'checks' module by extracting new module 'suites' and relevant functions, and add new pytest module for suites
